+++ lib/presentation/initial_page/widgets/page_data_widgets/trolley_running_mechanism_part_page.dart
@@ -6,7 +6,7 @@ import 'package:hmi_core/hmi_core_app_settings.dart';
 import 'package:hmi_core/hmi_core_translate.dart';
 import 'package:hmi_widgets/hmi_widgets.dart';
 ///
-/// Part of [InitialPage] - the trolley running mechanism.
+/// Part of [FirstPage] - the trolley running mechanism.
 /// FIve fields.
 class TrolleyRunningMechanismPartPage extends StatefulWidget {
   final SgocInit _fields;
